model and results plots code
Code to fit CITS models including overdispersion check, which then runs family selected (negative binomial as default, as per SAP)

generate counterfactual function to use to see what happens if CAB didn't happen, including summary table for statistics from model

there are two plot functions, one for generating standard cits plots comparing two groups, along with a save function in each (similar to the loess_graphs.R code in this subdirectory)

the other function generates a plot comparing the two slopes of the groups, allowing for a difference to be put at different lengths of time from the intervention. this allows for the comparison, for example, for only a few weeks after the intervention to up to 153.5 weeks (the maximum)

also includes the code to run the functions, including appending the data from the counterfactual function onto the list of information obtained from the cits, all the plotting functions, and also saving tidy model results and saving the summary tables from the counterfactual analysis.