add matching code check, update codelist
Update HIV code list after creating match checking code to find new codes based on the description field as the data are messy.

The match checking code checks for inclusion or exclusion (exclusions are not mandatory) keywords in the diagnostic description fields (10 in this case) and returns both a count of matched codes and a data frame. The distinct values in the data frame can be printed out so you can manually judge which codes you may be missing.